Remove chandrakkala anchor point
It is not really reqired. And if the chandrakkala anchor point
exist as mark class and no base anchor point found, in Mac systems
chandrakkala goes to relative to the x=zero of base.

That is, if p1 + xx rendered , xx appears at the beginning of p1.

In short, all mark class anchor points explicitly demands a base class
anchorpoint in such systems.